Sentence Parser for Stepheny
In C#, write a program that parses a sentence and replaces each word with the following: first letter, number of distinct characters between first and last character, and last letter.  For example, Smooth would become S3h.  Words are separated by spaces or non-alphabetic characters and these separators should be maintained in their original form and location in the answer.

The code must be syntactically correct and build in visual studio, either as a console or winforms application.

Please paste these instructions at the top of your completed assignment.